refactor: move `ResultDocstring`
